Headache refactor.
Also whoami.** and its kin  are all extern "C" know.

Currently whoami aka passport will dump Self User identifiers,  User groups,        privilages for the process selected to dump (itself as multitool/midas), and  if the same process is elvated or not.

There are explainations for priv, a notice if process is elevated along with a message and why  - just a string by follows the [!!!!!] convention Priv layed out